Forward port of r4596. Check env variables at startup to avoid a segfault in dataviewer.
Fix a segfault in dataviewer that happens with the LIGONDSIP variable
is not set. The updated script prints an error message and aborts.


git-svn-id: https://redoubt.ligo-wa.caltech.edu/svn/advLigoRTS/trunk@4599 6dcd42c9-f523-4c6d-aada-af552506706e

+++ src/dv/dataviewer.sh.cmake
@@ -11,6 +11,10 @@ else
 fi
 export GRACE_HOME
 export LANG=C
+if [ -z "${LIGONDSIP}" ]; then
+    echo "Your environment is not setup.  Plese set LIGONDSIP"
+    exit 1
+fi
 PATH=${PATH}:${GRACE_HOME}/bin
 $DVPATH/dv -s ${LIGONDSIP} -a $$ -b $DVPATH $*
 /bin/rm -f -r /tmp/$$DC
